+++ History.md
@@ -1,6 +1,7 @@
 0.4.0.pre
 ==========
 
+* Ensure file name for migration are valid
 * **BIG CHANGE** rename `Rails::Sequel` module as `SequelRails`, this becomes
   the namespace for all sequel-rails related classes.
 * Split `Rails::Sequel::Storage` class in multiple adapter for each db